Our first weapon: Splattershot Jr!
A lot of tricky things here, but we got there!! One notable decision was to split the upgraded version into two different powers, to ease the logic for replacing different weapons vs the same weapon. This'll be more annoying to maintain in certain ways, but I think it's the right call for simplicity!