Fix Fizzy Bomb & Killer Wail random target damage
Mm, the previous version without `multiDamage` always applied the right amount of damage for the *first* enemy in the room, even if the enemies had different ink amounts.

Now, we have the card compute `multiDamage`, and reverse-engineer the monster's index to know which entry to use.

I wonder if there's a more canonical way to do this? Are there seriously not official cards that do similar? I can't remember any offhand…