Tweak/nerf Splat Wall
Instead of the doubling hack, let's just double the number on the card, and the player can figure it out after using it the first time. Plus I do think it's just a smidge too strong right now, so a minor turn-down of the ink amount should help! (it's just. so wildly good with incoming multi-attacks. it's sooo good lol)