Reduce Booyah Bomb threshold
I just played a game with 2 Booyahs in the deck and never triggered it once lol

I don't mind the idea that it's not *easy* to get, but I'd like it to be like. not like a RARE occurrence, y'know?

Let's try 12 and see if it feels like too much